velky uklid na presenteru uzivatele
