+++ app/presenters/UzivatelPresenter.php
@@ -40,7 +40,6 @@ class UzivatelPresenter extends BasePresenter
     private $log;
     private $subnet;
     private $sloucenyUzivatel;
-    private $uzivatelskeKonto;
     private $prichoziPlatba;
     private $tabulkaUzivatelu;
     private $parameters;
@@ -50,7 +49,7 @@ class UzivatelPresenter extends BasePresenter
 
     /** @var Components\LogTableFactory @inject **/
     public $logTableFactory;
-    function __construct(Model\CryptoSluzba $cryptosvc, Model\PovoleneSMTP $alowedSMTP, Model\Parameters $parameters, Model\AccountActivation $accActivation, Model\UzivatelListGrid $ULGrid, Model\PrichoziPlatba $platba, Model\UzivatelskeKonto $konto, Model\SloucenyUzivatel $slUzivatel, Model\Subnet $subnet, Model\SpravceOblasti $prava, Model\CestneClenstviUzivatele $cc, Model\TypSpravceOblasti $typSpravce, Model\TypPravniFormyUzivatele $typPravniFormyUzivatele, Model\TypClenstvi $typClenstvi, Model\TypCestnehoClenstvi $typCestnehoClenstvi, Model\ZpusobPripojeni $zpusobPripojeni, Model\TechnologiePripojeni $technologiePripojeni, Model\Uzivatel $uzivatel, Model\IPAdresa $ipAdresa, Model\AP $ap, Model\TypZarizeni $typZarizeni, Model\Log $log) {
+    function __construct(Model\CryptoSluzba $cryptosvc, Model\PovoleneSMTP $alowedSMTP, Model\Parameters $parameters, Model\AccountActivation $accActivation, Model\UzivatelListGrid $ULGrid, Model\PrichoziPlatba $platba, Model\SloucenyUzivatel $slUzivatel, Model\Subnet $subnet, Model\SpravceOblasti $prava, Model\CestneClenstviUzivatele $cc, Model\TypSpravceOblasti $typSpravce, Model\TypPravniFormyUzivatele $typPravniFormyUzivatele, Model\TypClenstvi $typClenstvi, Model\TypCestnehoClenstvi $typCestnehoClenstvi, Model\ZpusobPripojeni $zpusobPripojeni, Model\TechnologiePripojeni $technologiePripojeni, Model\Uzivatel $uzivatel, Model\IPAdresa $ipAdresa, Model\AP $ap, Model\TypZarizeni $typZarizeni, Model\Log $log) {
         $this->cryptosvc = $cryptosvc;
         $this->spravceOblasti = $prava;
         $this->cestneClenstviUzivatele = $cc;
@@ -67,7 +66,6 @@ class UzivatelPresenter extends BasePresenter
         $this->log = $log;
         $this->subnet = $subnet;
         $this->sloucenyUzivatel = $slUzivatel;
-        $this->uzivatelskeKonto = $konto;
         $this->prichoziPlatba = $platba;
         $this->tabulkaUzivatelu = $ULGrid;
         $this->parameters = $parameters;
@@ -1434,112 +1432,4 @@ class UzivatelPresenter extends BasePresenter
     	return true;
     }
 
-    public function renderPlatba()
-    {
-        $id = $this->getParameter('id');
-        $pohyb = $this->uzivatelskeKonto->findPohyb(array('PrichoziPlatba_id' => intval($id), 'Uzivatel_id NOT' => null));
-        //\Tracy\Dumper::dump($pohyb->Uzivatel);
-        if($pohyb)
-        {
-            if($pohyb->Uzivatel_id)
-            {
-                $this->template->canViewOrEdit = $this->ap->canViewOrEditAP($this->uzivatel->getUzivatel($pohyb->Uzivatel_id)->Ap_id, $this->getUser());
-            }
-            else
-            {
-                $this->template->canViewOrEdit = $this->getUser()->isInRole('VV') || $this->getUser()->isInRole('TECH');
-            }
-            $this->template->u = $pohyb->Uzivatel;
-        }
-        else
-        {
-            $this->template->canViewOrEdit = false;
-            $this->template->u = null;
-        }
-
-        $this->template->canViewOrEditCU = $this->getUser()->isInRole('VV') || $this->getUser()->isInRole('TECH');
-        $this->template->canTransfer = $this->getUser()->isInRole('VV') || $this->getUser()->isInRole('TECH');
-        $this->template->p = $this->prichoziPlatba->getPrichoziPlatba($this->getParam('id'));
-    }
-
-    public function renderAccount()
-    {
-        $uzivatel = $this->uzivatel->getUzivatel($this->getParam('id'));
-
-        $this->template->canViewOrEdit = $this->ap->canViewOrEditAP($this->uzivatel->getUzivatel($this->getParam('id'))->Ap_id, $this->getUser());
-        $this->template->u = $uzivatel;
-
-        $stavUctuIn = $uzivatel->related('UzivatelskeKonto.Uzivatel_id')->where('castka>0')->sum('castka');
-        $this->template->sum_input = $stavUctuIn;
-        $stavUctuOut = $uzivatel->related('UzivatelskeKonto.Uzivatel_id')->where('castka<0')->sum('castka');
-        $this->template->sum_output = $stavUctuOut;
-    }
-
-    protected function createComponentAccountgrid($name)
-    {
-        $canViewOrEdit = false;
-    	$id = $this->getParameter('id');
-
-        //\Tracy\Dumper::dump($search);
-
-    	$grid = new \Grido\Grid($this, $name);
-    	$grid->translator->setLang('cs');
-        $grid->setExport('account_export');
-
-        if($id){
-            $seznamTransakci = $this->uzivatelskeKonto->getUzivatelskeKontoUzivatele($id);
-
-            $canViewOrEdit = $this->ap->canViewOrEditAP($this->uzivatel->getUzivatel($this->getParam('id'))->Ap_id, $this->getUser());
-
-        }
-
-        $grid->setModel($seznamTransakci);
-
-    	$grid->setDefaultPerPage(500);
-        $grid->setPerPageList(array(25, 50, 100, 250, 500, 1000));
-    	$grid->setDefaultSort(array('datum_cas' => 'DESC'));
-
-        $presenter = $this;
-        $grid->setRowCallback(function ($item, $tr) use ($presenter){
-                if($item->PrichoziPlatba_id)
-                {
-                    $tr->onclick = "window.location='".$presenter->link('Uzivatel:platba', array('id'=>$item->PrichoziPlatba_id))."'";
-                }
-                return $tr;
-            });
-
-        $grid->addColumnText('castka', 'Částka')->setSortable()->setFilterText();
-
-        $grid->addColumnDate('datum_cas', 'Datum')->setDateFormat(\Grido\Components\Columns\Date::FORMAT_DATETIME)->setSortable()->setFilterText();
-
-        $grid->addColumnText('TypPohybuNaUctu_id', 'Typ')->setCustomRender(function($item) {
-            return Html::el('span')
-                    ->alt($item->TypPohybuNaUctu_id)
-                    ->setTitle($item->TypPohybuNaUctu->text)
-                    ->setText($item->TypPohybuNaUctu->text)
-                    ->data("toggle", "tooltip")
-                    ->data("placement", "right");
-            })->setSortable();
-
-        $grid->addColumnText('poznamka', 'Poznámka')->setCustomRender(function($item){
-                $el = Html::el('span');
-                $el->title = $item->poznamka;
-                $el->setText(Strings::truncate($item->poznamka, 100, $append='…'));
-                return $el;
-                })->setSortable()->setFilterText();
-
-        $grid->addColumnText('PrichoziPlatba_id', 'Akce')->setCustomRender(function($item) use ($presenter)
-                {
-                    if ($item->PrichoziPlatba_id)
-                    {
-                        $uidLink = Html::el('a')
-                        ->href($presenter->link('Uzivatel:platba', array('id'=>$item->PrichoziPlatba_id)))
-                        ->title('Příchozí platba')
-                        ->setText('Příchozí platba');
-                        return $uidLink;
-                    } else {
-                        return ;
-                    }
-                });
-    }
 }
