+++ app/templates/UzivatelAccount/platba.latte
@@ -3,7 +3,7 @@
 
     {ifset $u}
     <h3>Příchozí platba pro: {if $canViewOrEdit}{$u->jmeno} {$u->prijmeni} {/if}({$u->nick}) <small> - {$u->Ap->Oblast->jmeno} {$u->Ap->jmeno}</small></h3>
-    <a href="{link Uzivatel:account id=>$u->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-arrow-up"></i> Zpět na přehled plateb</a>
+    <a href="{link UzivatelAccount:account id=>$u->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-arrow-up"></i> Zpět na přehled plateb</a>
     
     {if $canTransfer}
     <a href="{link SpravaNesparovanych:prevod id=>$p->id}" class="btn btn-default btn-sm" role="button"><i class="glyphicon glyphicon-usd"></i> Převést jinému uživateli (chybná platba)</a>
